+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@mnrendra/alias-resolver",
   "version": "2.2.0",
-  "description": "A utility to resolve alias paths.",
+  "description": "A utility for resolving alias paths.",
   "private": false,
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -26,7 +26,7 @@
   "homepage": "https://github.com/mnrendra/alias-resolver#readme",
   "dependencies": {
     "@mnrendra/types-aliases": "^1.1.0",
-    "acorn": "^8.12.1",
+    "acorn": "^8.14.0",
     "acorn-walk": "^8.3.4",
     "escodegen": "^2.1.0"
   },
@@ -34,8 +34,8 @@
     "@semantic-release/exec": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
     "@types/escodegen": "^0.0.10",
-    "@types/jest": "^29.5.13",
-    "@types/node": "^22.7.4",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.9.0",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "clean-package": "^2.2.0",
     "cz-conventional-changelog": "^3.3.0",
@@ -45,13 +45,13 @@
     "eslint-plugin-n": "^16.6.2",
     "eslint-plugin-promise": "^6.6.0",
     "jest": "^29.7.0",
-    "rollup": "^4.24.0",
+    "rollup": "^4.26.0",
     "rollup-plugin-dts": "^6.1.1",
     "rollup-plugin-esbuild": "^6.1.1",
-    "semantic-release": "^24.1.2",
+    "semantic-release": "^24.2.0",
     "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.3"
   },
   "keywords": [
     "alias",
