+++ src/types/index.ts
@@ -1,21 +1,4 @@
 export type {
-  Literal,
-  Program,
-  CallExpression,
-  VariableDeclaration,
-  ImportDeclaration
-} from 'acorn'
-
-export type {
-  SimpleVisitors
-} from 'acorn-walk'
-
-export type {
-  Aliases,
-  Alias
-} from '@mnrendra/types-aliases'
-
-export type {
   ResolveImport,
   ResolveRequire,
   ResolveDynamicImport
