+++ tests/index.ts
@@ -1,7 +1,9 @@
 import * as mocks from './mocks'
 import * as unmocks from './unmocks'
+import * as utils from './utils'
 
 export {
   mocks,
-  unmocks
+  unmocks,
+  utils
 }
