+++ tests/utils/index.ts
@@ -1,0 +1,5 @@
+import read from './read'
+
+export {
+  read
+}
